Move to standalone binary jaeger-v2
