+++ cmd/collector/main.go
@@ -28,7 +28,6 @@ import (
 
 	"github.com/jaegertracing/jaeger/cmd/collector/app"
 	"github.com/jaegertracing/jaeger/cmd/collector/app/flags"
-	"github.com/jaegertracing/jaeger/cmd/collector/internal/otelcmd"
 	"github.com/jaegertracing/jaeger/cmd/internal/docs"
 	"github.com/jaegertracing/jaeger/cmd/internal/env"
 	cmdFlags "github.com/jaegertracing/jaeger/cmd/internal/flags"
@@ -143,7 +142,6 @@ func main() {
 	command.AddCommand(env.Command())
 	command.AddCommand(docs.Command(v))
 	command.AddCommand(status.Command(v, ports.CollectorAdminHTTP))
-	command.AddCommand(otelcmd.Command())
 
 	config.AddFlags(
 		v,
