+++ cmd/jaeger-v2/.gitignore
@@ -1,0 +1,2 @@
+jaeger-v2-*-*
+jaeger-v2
